refactor: 로또 게임 구조 개선
- 로또 등수 그룹핑을 위한 Ranks 객체 작성

- LottoTicket 연산을 위한 LottoTickets 작성

+++ src/main/java/lotto/model/RankResult.java
@@ -6,9 +6,9 @@ public class RankResult {
 
     private final Rank rank;
 
-    private final int count;
+    private final long count;
 
-    public RankResult(Rank rank, int count) {
+    public RankResult(Rank rank, long count) {
         this.rank = rank;
         this.count = count;
     }
@@ -17,7 +17,7 @@ public class RankResult {
         return rank;
     }
 
-    public int getCount() {
+    public long getCount() {
         return count;
     }
